Split future_then_result into 2 files
To move the dependency to executors.
